Don't write directly to Ruby String object storage
Rather than overwriting part of the buffer that Ruby uses to store the
template String, create a temporary C string for use with mkdtemp.

In practice, there has never been a problem with this kind of direct
manipulation of Ruby String storage with any version of Ruby that I have
tested, but just to be future-proof we err on the side of caution here.

Signed-off-by: Wincent Colaiuta <win@wincent.com>